Harden the engine against bad files and failed rebuilds
Review pass over the three layers, each finding verified in code and
covered by a test where the core is involved.

Kotlin and JNI: the consumer rule keeps the NativeEngine class name, not
only its members; release() queues the destruction behind a running
decode instead of blocking the caller; loads after release are no-ops;
the any-thread readers lock against destroy; the JNI bridge checks for a
pending exception and logs an invalid Surface.

Engine: a swapchain or pipeline rebuild that fails drops the surface
instead of drawing through a null pipeline; a sort result for the old
world is discarded when a new one is uploaded (it indexed past a smaller
splat buffer); the level of detail budget is handed to the render thread
with the world instead of being read across threads; the upload checks
every Vulkan result; a teleport publishes the pose at once.

Core: the SPZ gzip container is inflated here with a real ceiling instead
of trusting the trailer; GLB accessors that overflow the bounds check,
miss a field or carry short arrays are rejected; a far vertex grows the
collider cells instead of the grid; non-finite rotations are rejected;
a level of detail node can hold more than 65k children.

+++ packages/splat-core/src/lod/LodTree.cpp
@@ -203,7 +203,7 @@ LodTree buildLodTree(SplatCloud cloud, const LodBuildOptions& options) {
     const uint32_t old = order[head];
     const auto& kids = nodes.children[old];
     tree.layout[head].childStart = static_cast<uint32_t>(order.size());
-    tree.layout[head].childCount = static_cast<uint16_t>(kids.size());
+    tree.layout[head].childCount = static_cast<uint32_t>(kids.size());
     for (uint32_t kid : kids) order.push_back(kid);
   }
 
@@ -278,7 +278,7 @@ void selectLodNodes(const LodTree& tree, Vec3 origin, const LodView& view, std::
       continue;
     }
     const uint32_t node = bucket.back();
-    const uint16_t kids = layout[node].childCount;
+    const uint32_t kids = layout[node].childCount;
     if (kids == 0) {
       bucket.pop_back();
       out.push_back(node);
